refactor: 페이징 API 응답 시 다음 페이지 여부 추가
